Move tooltip to place of usage
