+++ front/app/components/admin/ImageCropper/messages.ts
@@ -10,18 +10,4 @@ export default defineMessages({
     id: 'app.components.Admin.ImageCropper.infoLinkText',
     defaultMessage: 'recommended ratio',
   },
-  imageSupportPageURL: {
-    id: 'app.components.Admin.ImageCropper.imageSupportPageURL',
-    defaultMessage:
-      'https://support.citizenlab.co/en/articles/1346397-what-are-the-recommended-dimensions-and-sizes-of-the-platform-images',
-  },
-  imageTooltip: {
-    id: 'app.components.Admin.ImageCropper.imageTooltip',
-    defaultMessage:
-      'For more information on recommended image resolutions, {supportPageLink}.',
-  },
-  imageSupportPageText: {
-    id: 'app.components.Admin.ImageCropper.imageSupportPageText',
-    defaultMessage: 'visit our support center',
-  },
 });
