+++ front/app/containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/BannerImageFields/messages.ts
@@ -1,0 +1,18 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  imageSupportPageURL: {
+    id: 'app.components.Admin.ImageCropper.imageSupportPageURL',
+    defaultMessage:
+      'https://support.citizenlab.co/en/articles/1346397-what-are-the-recommended-dimensions-and-sizes-of-the-platform-images',
+  },
+  imageTooltip: {
+    id: 'app.components.Admin.ImageCropper.imageTooltip',
+    defaultMessage:
+      'For more information on recommended image resolutions, {supportPageLink}.',
+  },
+  imageSupportPageText: {
+    id: 'app.components.Admin.ImageCropper.imageSupportPageText',
+    defaultMessage: 'visit our support center',
+  },
+});
